chore(hybridcloud) Move unguarded_write out of testutils (#52807)
When opening the mulligan pull requests, I remembered that I'll also
need to move the unguarded_write function out of testutils first, as I
need to update getsentry before sentry.
