+++ src/sentry/silo/__init__.py
@@ -1,1 +1,2 @@
 from .base import *  # NOQA
+from .safety import *  # NOQA
